move DocumentFormatter to formatter.py; implement TaskEncodingFormatter
